[Metal] Obsolete and fix MTLDevice.CreateBufferNoCopy.
MTLDevice.CreateBufferNoCopy must be called with memory allocated with
vm_allocate or mmap, which will hardly be the case for managed arrays, so
obsolete this overload.

Still, the method might never work, but in the case it actually does for some
reason, don't free the handle to the managed memory until the MTLDevice is
done with it.