[generator] Don't print multiple EditorBrowsable attributes. Fixes xamarin/maccore#1661. (#6179)
It's not allowed to have multiple EditorBrowsable attributes on the same
member, so make sure the generatod doesn't generate such code.

We might want to print an "[EditorBrowsable (Never)]" because a particular API
is obsolete (i.e. has the "[Obsolete]" attribute), but if the API also has an
EditorBrowsable attribute already, use only that one instead.

Fixes https://github.com/xamarin/maccore/issues/1661.