[system-dependencies] Fix xcode-provisioning extraction
This should fix:

```
        Extracting /tmp/x-provisioning/Xcode_11_Beta_2.xip...
        Installing Xcode 11.0 to /Applications/Xcode11-beta2.app...
mv: rename *.app to /Applications/Xcode11-beta2.app: No such file or directory
```

This happened because we're copying the .xip to /tmp/x-provisioning, then try to extract it but for some reason that extraction always end up in ~/Downloads.
Therefore the move fails because the extracted .app is not in /tmp/x-provisioning.
Note: Apple's Xcode .xip are always extracted with a name of Xcode[-beta].app so hardcode that.
I don't know if it's because the original .xip was downloaded to ~/Download or if macOS's archive utility is just ignoring its preferences (save expended files into same directory as archive).
I couldn't find the option to force the archiver to extract to a specific location (feedback welcome).