[system-dependencies] Use 'xip' to decompress xip files.
Archive Utility compresses into a directory specified by one of its
preferences, so we can't really say for sure where it will put the extracted
xip.

This is of course all sorts of fun, except none of the good ones.

So use the command-line xip utility to extract the xip instead, which is
documented to extract into the current directory.