[generator] Reference bound *Appearance types using global:: syntax. Fixes #6834.
The generator will create special *Appearance types (these are nested
classes). If we've bound a type with the same *Appearance name, we can end up
in a situation where the csc compiler uses the the type we don't want due to
C#'s resolution rules - this happens if the bound *Appearance type is
referenced from the containing type of the special *Appearance type. So always
reference the bound *Appearance types using global:: syntax.

Fixes https://github.com/xamarin/xamarin-macios/issues/6834.