[coreaudio] Update for Xcode 11 GM
With xcode11 GM the tvOS and macOS headers drifted from the iOS ones.
It's not clear if they should be different (hopefully not) or if iOS
has been forgotten...

WHile waiting for Apple's feedback (and the avoid breaking changes)
the new (conflicting) enum values are not included.

Internal reference: https://github.com/xamarin/maccore/issues/1960