[coreimage] Update for Xcode 11[.1]
```
Incomplete (draft) - it's missing a bunch of `[Obsolete]` but those are
not breaking changes so the _core_ can be reviewed.

Pay attention to naming (many conflicts) and comment if you have better
names suggestions.

It's also a big PR with a lot of copy/paste so watch out for typos!
```

Apple decided to expose most (but not all) `CIFilter` using protocols
(instead of weakly named dictionaries). Most of this maps well with
our strong bindings but there are cases where we:

* missing some properties (easy, there were added); or
* used a different types [1] and new members were required

[1] Often ours are better (using `float` for a `bool` is not optimal)
but we do not have `[BindAs]` for protocols :( to _fix_ them