Drop the Xcode 9.4 dependency. (#7044) (#7135)
* Drop the Xcode 9.4 dependency. (#7044)

* Drop the Xcode 9.4 dependency.

Also bump mono to get the removal of the mac32 binaries.

New commits in mono/mono:

* mono/mono@beb9a1b182a [sdks] Remove the mac32 build.
* mono/mono@747a919a06e [ci] Make ios/mac sdks archive URL more predictable
* mono/mono@114013096e1 [ci] Build iOS/Mac Mono sdks archive using Xcode 11
* mono/mono@10a24f3ea1d Implement WriteCore and ReadCore in DeflateStream
* mono/mono@a925846b1f0 [offsets-tool] Install clang into the user-specific python directory. (#16933)
* mono/mono@fe64a4765e6 [2019-06] Bump msbuild and sdk versions to 3.0.1xx latest (#16870)
* mono/mono@7293597b905 [corlib] Fix building nunit-lite twice (#16910)
* mono/mono@1648e886873 Rename bundle identifier for the various Mono.frameworks we create for Xamarin.iOS. Fixes xamarin/xamarin-macios#7005. (#16896)
* mono/mono@a6b5187d76a [metadata] Fix leaks when handling a few attributes (#16675) (#16851)
* mono/mono@7da9a041b3b [2019-06] Bump to mono/corefx@e79cf5b
* mono/mono@2b7050bdf36 [2019-06] Add RenamedEvent* to FSW sources from CoreFX (#16758)
* mono/mono@4f5ed502c6e [msbuild] pick up p4 versions
* mono/mono@f04ee2219d5 [2019-06][msbuid][roslyn] Bump msbuild and roslyn-binaries to pick up dotnet 3.0.100-p9 toolset
* mono/mono@6b4b99e571b Vtable [i] can be null so this should be check before use it. Fixes #16712

Diff: https://github.com/mono/mono/compare/7af64d1ebe9e9ee305cdae8ec5995c9521cbcf19..beb9a1b182a14986e836864e5d555c3b5ec52ba0

* [tests] Add a fat macOS dylib for testing purposes.

Add a binary version of a fat macOS dylib (because we can't create one when we
need it since we can't create 32-bit slice anymore).

It was created like this (in tests/test-libraries):

	$ cat test.m
	int theUltimateAnswer ()
	{
		return 42;
	}

	$ /Applications/Xcode94.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/clang test.m -olibtest.i386.dylib -shared -isysroot /Applications/Xcode94.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX.sdk -framework Foundation -framework CoreLocation -lz  -arch i386
	$ lipo -create libtest.i386.dylib .libs/macos/libtest.dylib -output libtest-fat.dylib

* [tests] Adjust XM tests to XM not having fat dylibs anymore.

* [tests] Adjust product tests to some libraries not being fat anymore.

* [tests] Don't treat an Xcode with the same major version number as old.

Fixes an issue in the MT0091 test, where it would fail on tvOS because the
test wanted to use an older Xcode, and we could end up returning Xcode 11.0
when the current Xcode is 11.1. Since the test depends on using the OS SDK as
it was designed for (technically using an OS SDK earlier than the latest), it
ended up failing because while the iOS SDK was bumped in Xcode 11.1, the tvOS
SDK was not.