[coreimage] Do the right thing wrt 'inputImage'
Not all filters support `inputImage` but we exposed it everywhere with
the `Image` property in `CIFilter`.

That cause problems with filters that did not support it, e.g. crashing
in the debugger, so we had to check it's existence before [get/set]ing
it. That's not good for performance (now a bit optimized) and the API
actually makes it harder to discover which filters really support it.

This also solve _part of_ the warnings coming from the static registrar