[Networking] Add a generic method to get the metadata from the context.
There are different types of NWProtocolMetadata, being one of them (to be
added later) the NWFramerMessage that exposes extra methods.

With the current API we can only do:

var metadata = context.GetProtocolMetadata (protocol.ProtocolDefinition);

Which just returns the generic object, and if the protocol definition is
from a NWFramer, we want to be able to get the correct metadata object.

With the provided method you can do:

var metadata = context.GetProtocolMetadata<NWFramerMessage> (protocol.ProtocolDefinition);

Which allows the user to specify the expected metadata type from a given
protocol definition.