[Network] Add more convinient callbacks. (#7252)
Providing callbacks with DispatchData and Int + length is not nice for
the non advance users. Added set of methods that will use callbacks that
will provide the users with ReadOnlySpan, which is nicer and makes it
work with the API easier.

Also, the change allows users to avoid unsafe code.